+++ Compiler/expressions/Bind.ts
@@ -4,9 +4,7 @@
  */
 
 import { BindExpressionVisitor } from 'Compiler/codegen/Expression';
-import { EventNode } from './EventNode';
-import { EventChain } from './EventChain';
-import { IAttributeValue } from './IAttributeValue';
+import { EventNode, EventChain, IAttributeValue } from './Event';
 import * as FSC from 'Compiler/modules/data/utils/functionStringCreator';
 
 /**
