+++ Compiler/expressions/Event.ts
@@ -4,15 +4,106 @@
  */
 
 import { EventExpressionVisitor } from 'Compiler/codegen/Expression';
-import { EventNode } from './EventNode';
-import { EventChain } from './EventChain';
-import { IAttributeValue } from './IAttributeValue';
 import * as FSC from 'Compiler/modules/data/utils/functionStringCreator';
+import { VariableNode } from 'Compiler/expressions/Statement';
+
+/**
+ * Интерфейст значения атрибута.
+ * @todo Отрефакторить парсеры и перенести интерфейс в определения Wasaby-узлов.
+ */
+export interface IAttributeValue {
+   /**
+    * Значение атрибута можеть быть "склейкой" разных узлов - локализация, текст, выражение.
+    */
+   data: VariableNode[];
+   /**
+    * Флаг, необходимый для различия опций контрола от атрибутов.
+    * @todo Избыточный флаг. Отрефакторить код и избавиться от него.
+    */
+   property: boolean;
+   /**
+    * Тип атрибута.
+    * @todo Избыточный флаг. Отрефакторить код и избавиться от него.
+    */
+   type: string;
+}
 
 /**
  * Паттерн обработчика события для имени атрибута.
  */
 const EVENT_NAME_PATTERN = /^(on:[A-z0-9])\w*$/;
+
+/**
+ * Класс цепочки обработчиков. Содержит коллекцию узлов EventNode для конкретного события.
+ *
+ * ВАЖНО: в начале идут bind-обработчики, а затем - события.
+ */
+export class EventChain extends Array<EventNode> {
+   /**
+    * Флаг, по которому различается EventChain от обычного массива.
+    * @todo Избыточный флаг. Отрефакторить код и избавиться от него.
+    */
+   readonly events: boolean = true;
+
+   /**
+    * Подготовить цепочку обработчиков: создать массив или вернуть имеющийся.
+    * @param originChain {EventChain} Исходная цепочка обработчиков для конкретного события либо undefined.
+    */
+   static prepareEventChain: (originChain?: EventChain) => EventChain;
+}
+
+EventChain.prepareEventChain = function prepareEventChain(originChain?: EventChain): EventChain {
+   if (!originChain) {
+      return new EventChain();
+   }
+   return originChain;
+};
+
+interface IEventNodeCfg {
+   args?: string;
+   value: string;
+   viewController: string;
+   data?: string;
+   handler: string;
+   isControl: boolean;
+   context?: string;
+}
+
+/**
+ * Данный класс представляет узел обработчика события.
+ */
+export class EventNode {
+   /**
+    * Аргументы, переданные функции-обработчику события.
+    */
+   args: string;
+   /**
+    * Имя функции-обработчика события.
+    */
+   value: string;
+
+   viewController: string;
+   data: string;
+   handler: string;
+   isControl: boolean;
+   context: string;
+
+   bindValue: string;
+
+   /**
+    * Инициализировать новый узел.
+    * @param cfg {IEventNodeCfg}
+    */
+   constructor(cfg: IEventNodeCfg) {
+      this.args = cfg.args;
+      this.value = cfg.value;
+      this.viewController = cfg.viewController;
+      this.data = cfg.data;
+      this.handler = cfg.handler;
+      this.isControl = cfg.isControl;
+      this.context = cfg.context;
+   }
+}
 
 /**
  * Создать узел обработчика события по атрибуту.
